Move ChatHistory to the separate component. Separate logic and fix some bugs
